requirements.txt
langchain
langchain-openai
langchain-community
langchainhub
langchain-core
python-dotenv
black
ipykernel
flask
flask-Cors

